+++ src/TaskEither.ts
@@ -18,10 +18,10 @@ import { bindTo_, bind_, flow, identity, Lazy, pipe, Predicate, Refinement } fro
 import { Functor2 } from './Functor'
 import { IO } from './IO'
 import { IOEither } from './IOEither'
-import { Monad2, Monad2C } from './Monad'
+import { Monad2 } from './Monad'
 import { MonadIO2 } from './MonadIO'
-import { MonadTask2, MonadTask2C } from './MonadTask'
-import { MonadThrow2, MonadThrow2C } from './MonadThrow'
+import { MonadTask2 } from './MonadTask'
+import { MonadThrow2 } from './MonadThrow'
 import { Monoid } from './Monoid'
 import { Option } from './Option'
 import { Semigroup } from './Semigroup'
@@ -647,32 +647,6 @@ export function getAltTaskValidation<E>(SE: Semigroup<E>): Alt2C<URI, E> {
               )
         )
       )
-  }
-}
-
-// TODO: remove in v3
-/**
- * @category instances
- * @since 2.0.0
- */
-export function getTaskValidation<E>(
-  SE: Semigroup<E>
-): Monad2C<URI, E> & Bifunctor2<URI> & Alt2C<URI, E> & MonadTask2C<URI, E> & MonadThrow2C<URI, E> {
-  const applicativeTaskValidation = getApplicativeTaskValidation(T.ApplicativePar, SE)
-  const altTaskValidation = getAltTaskValidation(SE)
-  return {
-    URI,
-    _E: undefined as any,
-    map: map_,
-    ap: applicativeTaskValidation.ap,
-    of,
-    chain: chain_,
-    bimap: bimap_,
-    mapLeft: mapLeft_,
-    alt: altTaskValidation.alt,
-    fromIO,
-    fromTask,
-    throwError
   }
 }